- Add Tests for parsing and serializing DateTime
