+++ test/test_dartson.dart
@@ -61,7 +61,6 @@ void main() {
     var test = new TestClass1();
     test.name = "test1";
     String str = serialize(test);
-    print(str);
     expect(str,'{"name":"test1"}');
   });
   
@@ -87,16 +86,7 @@ void main() {
       '{"name":"test","list":[1,2,3],"getter":{"name":"get it"}}');
   });
 
-  test('parse: generic map', () {
-    print('begin parsing generic map');
-    MapClass test = parse('{"mapOfSimpleClass": {"test": {"name": "test"}, "test2": {"name": "test2"}}}', MapClass);
-
-    expect(test.mapOfSimpleClass["test"].name, "test");    
-    expect(test.mapOfSimpleClass["test2"].name, "test2");
-  });
-  
   test('parse: parser simple', () {
-    print('begin parsing simple');
     TestClass1 test = parse('{"name":"test","matter":true,"intNumber":2,"number":5,"list":[1,2,3],"map":{"k":"o"},"the_renamed":"test"}', TestClass1);
     expect(test.name, 'test');
     expect(test.matter, true);
@@ -142,7 +132,13 @@ void main() {
     expect(list.list[0], 1);
   });
   
-  
+  test('parse: generic map', () {
+    MapClass test = parse('{"map": {"test": {"name": "test"}, "test2": {"name": "test2"}}}', MapClass);
+
+    expect(test.map["test"].name, "test");
+    expect(test.map["test2"].name, "test2");
+  });
+
   test('parse: simple map', () {
     SimpleMap test = parse('{"map": {"test": "test", "test2": "test2"}}', SimpleMap);
 
@@ -189,6 +185,22 @@ void main() {
       registerTransformer(new SimpleTransformer<DateTime>());
       expect(hasTransformer(DateTime), true);
     });
+
+    test('parse: DateTime', () {
+      var date = new DateTime.now();
+      var ctg = parse('{"testDate":"${date.toString()}"}', SimpleDateContainer);
+      expect(ctg.testDate is DateTime, true);
+      expect(ctg.testDate == date, true);
+    });
+
+    test('serialize: DateTime', () {
+      var obj = new SimpleDateContainer();
+      obj.testDate = new DateTime.now();
+      var str = serialize(obj);
+      expect(str, '{"testDate":"${obj.testDate.toIso8601String()}"}');
+    });
+
+
 }
 
 class SimpleTransformer<T> extends TypeTransformer {
@@ -261,7 +273,7 @@ class ListClass {
 }
 
 class MapClass {
-  Map<String, SimpleClass> mapOfSimpleClass;
+  Map<String, SimpleClass> map;
 }
 
 class SimpleList {
